Loading files and some attempts in finding arrival events
Note that you have to change the path to your files. Currently also just for one day.

Finding an arrival event turned out to be figuring out how to store dates and times in the right format (so that we can do arithmetic on these values). However, Python is playing hard to get on me and it doesn't really seem to work.